Fixing un-manage flow issue when cluster import fails
tendrl-bug-id: Tendrl/monitoring-integration#436

+++ tendrl/monitoring_integration/flows/delete_monitoring_data/__init__.py
@@ -64,23 +64,27 @@ class DeleteMonitoringData(flows.BaseFlow):
                 graphite_utils.get_data_dir_path(),
                 integration_id
             )
-        try:
-            shutil.move(resource_path, archive_path)
-        except Exception as ex:
-            raise FlowExecutionFailedError(
-                "Failed to archive the monitoring data. Error: (%s)" %
-                ex
-            )
+        # if source path not exist then no need to archive and raise
+        # exception, because it will affect when user try to do
+        # unmanage flow after import flow fail
+        if os.path.exists(resource_path):
+            try:
+                shutil.move(resource_path, archive_path)
+            except Exception as ex:
+                raise FlowExecutionFailedError(
+                    "Failed to archive the monitoring data. Error: (%s)" %
+                    ex
+                )
 
-        # Log an event mentioning the archive data location
-        logger.log(
-            "debug",
-            NS.publisher_id,
-            {
-                "message": "%s un-managed.\n"
-                "Archived monitoring data to %s" %
-                (integration_id, archive_path)
-            }
-        )
+            # Log an event mentioning the archive data location
+            logger.log(
+                "debug",
+                NS.publisher_id,
+                {
+                    "message": "%s un-managed.\n"
+                    "Archived monitoring data to %s" %
+                    (integration_id, archive_path)
+                }
+            )
 
         return True
